ddl: add a ddl_op file
